WSB_portal/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
plotly==5.23.0
numpy==2.3.5
psycopg[binary]==3.2.12
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-jose[cryptography]==3.3.0
redis==5.2.0
APScheduler==3.10.4
slowapi==0.1.9
openpyxl==3.1.5

